x/did/types/genesis.go
package types

import (
	"fmt"
)

// DefaultIndex is the default global index
const DefaultIndex uint64 = 1

// DefaultGenesis returns the default genesis state
func DefaultGenesis() *GenesisState {
	return &GenesisState{
		OrgsList: []Orgs{},
		// this line is used by starport scaffolding # genesis/types/default
		Params: DefaultParams(),
	}
}

// Validate performs basic genesis state validation returning an error upon any
// failure.
func (gs GenesisState) Validate() error {
	// Check for duplicated ID in orgs
	orgsIdMap := make(map[uint64]bool)
	orgsCount := gs.GetOrgsCount()
	for _, elem := range gs.OrgsList {
		if _, ok := orgsIdMap[elem.Id]; ok {
			return fmt.Errorf("duplicated id for orgs")
		}
		if elem.Id >= orgsCount {
			return fmt.Errorf("orgs id should be lower or equal than the last id")
		}
		orgsIdMap[elem.Id] = true
	}
	// this line is used by starport scaffolding # genesis/types/validate

	return gs.Params.Validate()
}
